hhli/requirements.txt
Django==1.9
django-crispy-forms==1.5.2
django-filter==0.11.0
django-material==0.5.1
djangorestframework==3.3.2
drfdocs==0.0.5
Markdown==2.6.5
